[TP-104] Refactor : BaseEntity에서 BaseTimeEntity 분리
